+++ HISTORY.md
@@ -6,6 +6,7 @@
 - Fix: Clean whitespace in HTML attributes. Thanks https://github.com/grantnorwood
 - Add: Add the ability to move categories by tag. Thanks https://github.com/grantnorwood
 - Tweak: Change the capability to `edit_others_posts`. (issue #7)
+- Tweak: Make Bulk_Move class as a singleton
 
 ### v1.1.1 (2014-02-12) (Dev time: 0.5 hours)
 - Add: Add German translation